refactor(classification): split leaf classifier into classification_core (#499 Cluster B)
Extract ClassificationResult / TYPE_PRIORS / classify_sentence /
get_source_adjusted_prior into the leaf module classification_core
(imports only correction + models). derivation.py now imports from
classification_core, breaking the (derivation, classification) edge
that closed two cycles flagged by CodeQL on main:

  - classification ↔ derivation ↔ derivation_worker
  - classification ↔ derivation ↔ derivation_worker ↔ scanner

classification.py keeps the orchestration layer (start_onboard_session,
accept_classifications, HostClassification dataclass) and re-exports
the leaf symbols from classification_core for backward compat. The
function-local imports inside the orchestration functions stay; they
no longer close a cycle because derivation no longer reaches back to
classification.

No behavior change. 2923 tests pass.
